Problem: Supervision tree is constructed in a way, that not follows the domain logic
Solution:
 - We should keep closely connected processes under the same supervisor,
   this would allow us to avoid inconsistent state in case of failure.
   For example, if event handler fails, we want dependent read models
   (that filled up by this event handler) to be restarted as well.
