+++ lib/zhr_devs/application.ex
@@ -2,10 +2,12 @@ defmodule ZhrDevs.Application do
   @moduledoc false
   use Application
 
-  alias ZhrDevs.Otp.EventHandlersSupervisor
-  alias ZhrDevs.Otp.ProcessManagersSupervisor
-  alias ZhrDevs.Otp.ProjectionsSupervisor
-  alias ZhrDevs.Otp.SubmissionSupervisor
+  alias ZhrDevs.Otp.SubmissionCheckSupervisor
+  alias ZhrDevs.Otp.SubmissionsSupervisor
+  alias ZhrDevs.Otp.TasksSupervisor
+  alias ZhrDevs.Otp.TransactionalEventsHandler
+
+  alias ZhrDevs.IdentityManagement
 
   @impl true
   def start(_type, _args) do
@@ -14,10 +16,11 @@ defmodule ZhrDevs.Application do
       {DynamicSupervisor, strategy: :one_for_one, name: ZhrDevs.DynamicSupervisor},
       {Task.Supervisor, name: ZhrDevs.EmailsSendingSupervisor},
       ZhrDevs.App,
-      ProjectionsSupervisor,
-      ProcessManagersSupervisor,
-      SubmissionSupervisor,
-      EventHandlersSupervisor,
+      IdentityManagement.EventHandler,
+      TasksSupervisor,
+      SubmissionsSupervisor,
+      SubmissionCheckSupervisor,
+      TransactionalEventsHandler,
       cowboy_child_spec()
     ]
